Move Router to separate file
